Use relative paths in Notice component
This is necessary because we are importing this component in
an external environment (wp-calypso) that doesn't know how
to access to `src/state/constants`.

This may be set up as absolute (to this repo) paths when
the external API for happychat-client is complete -
at this point we are importing components and need them to
be isolated.