Change when the ongoingConversation event is fired
Some clients hide the chat form status if the availability is false,
but they don't want to hide it if there is an ongoingConversation.

When an operator has 1 spot left, we're going to call subscribers
with availability=false and ongoingConversation=true.
But because the availability event happens before the chatStatus is "assigned",
we need to consider the "assigning" chatStatus as an ongoingConversation,
to prevent the race condition.