Fill the new iframe with loading indicator
By creating the hr.spinner-line when the iframe is created,
the perceived performance is a bit better.

When React is initialized it is going to use that same node
to render the content, so that'll be overwritten. By the way,
the same spinner line is the first thing React is going to render
when it takes over the targetNode, so the user won't see anything.