Revert use of lodash-webpack-plugin
After a few days using it, I've ran into some bugs. Some of them are
obvious and easily fixable (such as not using the object property name
to destructure some params), another are very subtle and could be easily
missed and then found at run-time (such as the startsWith method
failing when the first value is a number, not a string).

This optimization will require more thinking to be ready to ship.