Remove unused message type-switching
This was an early optimization that's totally unnecessary right now. It looked at the first message in any message grouping and chose how to render based on that first message's type. But we only have "message" types in the system right now. If we add more in the future, the type we add and the way we want it to render will inform the way the code should be structured.