Clean up message grouping
The messages were grouped using a `reduce` function, which can feel like the "right" thing to do, but I almost always find plain old `forEach` to more readable.

The rendering was also singling out the first message in the group, but it renders the exact same as the others. So I removed that unneeded separation.