Clean up Timeline component (#336)
* Remove properties that are not in use

* Remove Emojify/twemoji

These are already being removed in Calypso: https://github.com/Automattic/wp-calypso/pull/56573

They were also not applied consistently — messages with links were not emojified.

* Render all messages with the same function

There's no need to split rendering of "messages with links" from "messages without links" when they share so much in common. Turns out, the "with links" function rendered "without links" messages perfectly fine, so I consolidated to that one and renamed it `renderMessage()`.

* Remove unused message type-switching

This was an early optimization that's totally unnecessary right now. It looked at the first message in any message grouping and chose how to render based on that first message's type. But we only have "message" types in the system right now. If we add more in the future, the type we add and the way we want it to render will inform the way the code should be structured.

* Clean up message grouping

The messages were grouped using a `reduce` function, which can feel like the "right" thing to do, but I almost always find plain old `forEach` to more readable.

The rendering was also singling out the first message in the group, but it renders the exact same as the others. So I removed that unneeded separation.

* Remove functional timeline conditionals

Instead of routing rendering through multiple functional conditional checks, let's do it the old fashioned way — `if` statements. This also lets us choose which props to send into each rendered component (though for now, I decided against tweaking that too much).

* Consolidate renderWelcomeMessage right into the Timeline component

* Decouple message rendering from message formatting

Splits out the function that formats a message, from the function that creates the message component wrapper and metadata (like "edited" flags).